(maint) Use beaker-hostgenerator to create nodesets
This pulls in beaker-hostgenerator and adds a new rake task
`gen_nodeset` for utilizing it. The output of the rake task on success
is the location of the generated nodeset for use when calling Beaker
later on.

Example workflow:
```
export BEAKER_nodeset=`bundle exec rake gen_nodeset`
bundle exec rake beaker
```